feat(canary-web-components): refactor calendar in ic-date-picker into a standalone component
Create new ic-calendar component, refactor code in ic-date-picker so that it now uses the new
calendar component, ensure ic-date-picker still functions as expected, update tests and add new
stories for ic-calendar

. #3357
