+++ packages/canary-web-components/src/components/ic-calendar/ic-calendar.mdx
@@ -1,0 +1,6 @@
+import { Meta, Stories } from "@storybook/blocks";
+import * as IcCalendarStories from "./ic-calendar.stories";
+
+<Meta of={IcCalendarStories} />
+
+<Stories />
