+++ packages/canary-web-components/src/components/ic-calendar/test/a11y/ic-calendar.test.a11y.ts
@@ -1,0 +1,13 @@
+import { fixture } from "@open-wc/testing-helpers";
+import { axe } from "jest-axe";
+import { checkShadowElementRendersCorrectly } from "../../../../utils/testa11y.helpers";
+
+describe("ic-calendar", () => {
+  it("passes accessibility", async () => {
+    const el = await fixture(`
+    <ic-calendar></ic-calendar>
+    `);
+    checkShadowElementRendersCorrectly(el);
+    expect(await axe(el)).toHaveNoViolations();
+  });
+});
